Serparting components
